Adding START_XVFB as a env var to start (or not) Xvfb, in case a user wants to run Chrome and Firefox in headless mode and thinks that Xvfb uses too many resources. Fixes #429 and fixes #567.

+++ NodeBase/start-xvfb.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
-export GEOMETRY="${SCREEN_WIDTH}""x""${SCREEN_HEIGHT}""x""${SCREEN_DEPTH}"
+if [ "${START_XVFB}" = true ] ; then
+  export GEOMETRY="${SCREEN_WIDTH}""x""${SCREEN_HEIGHT}""x""${SCREEN_DEPTH}"
 
-rm -f /tmp/.X*lock
+  rm -f /tmp/.X*lock
 
-/usr/bin/Xvfb ${DISPLAY} -screen 0 ${GEOMETRY} -ac +extension RANDR
+  /usr/bin/Xvfb ${DISPLAY} -screen 0 ${GEOMETRY} -ac +extension RANDR
+else
+  echo "Xvfb won't start. Chrome/Firefox can only run in headless mode. Remember to set the 'headless' flag in your test."
+fi
